requirements.txt
click==7.1.2
colorama==0.4.4
Flask==1.1.1
Flask-Cors==3.0.8
importlib-metadata==1.4.0
itsdangerous==1.1.0
Jinja2==2.11.0
MarkupSafe==1.1.1
six==1.14.0
Werkzeug==0.16.1
zipp==2.0.1
pytest==6.2.2
gunicorn==20.0.4
pandas~=1.3.0
cassandra-driver~=3.25.0
scikit-learn~=0.22.1
numpy~=1.21.6
kneed~=0.5.1
xgboost~=0.90